+++ packages/cheqd-blockchain-modules/jest.config.js
@@ -3,18 +3,10 @@ export default {
   moduleNameMapper: {
     "^@docknetwork/credential-sdk/(.*)$":
       "<rootDir>/../credential-sdk/dist/esm/$1",
-    "^@docknetwork/dock-blockchain-api/(.*)$":
-      "<rootDir>/../dock-blockchain-api/dist/esm/$1",
     "^@docknetwork/cheqd-blockchain-api/(.*)$":
       "<rootDir>/../cheqd-blockchain-api/dist/esm/$1",
-    "^@docknetwork/dock-blockchain-modules/(.*)$":
-      "<rootDir>/../dock-blockchain-modules/dist/esm/$1",
-    "^@docknetwork/dock-blockchain-api$":
-      "<rootDir>/../dock-blockchain-api/dist/esm/index.js",
     "^@docknetwork/cheqd-blockchain-api$":
       "<rootDir>/../cheqd-blockchain-api/dist/esm/index.js",
-    "^@docknetwork/dock-blockchain-modules$":
-      "<rootDir>/../dock-blockchain-modules/dist/esm/index.js",
     "^@cheqd/sdk(.*)$": "<rootDir>/../../node_modules/@cheqd/sdk/build/cjs/$1",
     "^uint8arrays$": "<rootDir>/../../node_modules/uint8arrays/dist/src",
     "^file-type$": "<rootDir>/../../node_modules/file-type/index.js",