+++ packages/dock-blockchain-api/src/index.js
@@ -1,3 +1,0 @@
-import DockAPI from './api';
-
-export { DockAPI };