+++ packages/dock-blockchain-modules/jest.config.js
@@ -6,12 +6,10 @@ export default {
     "^@docknetwork/credential-sdk/(.*)$": resolve(
       "../credential-sdk/dist/esm/$1"
     ),
-    "^@docknetwork/dock-blockchain-api/(.*)$": resolve(
-      "../dock-blockchain-api/dist/esm/$1"
-    ),
-    "^@docknetwork/dock-blockchain-api$": resolve(
-      "../dock-blockchain-api/dist/esm/index.js"
-    ),
+    "^@docknetwork/dock-blockchain-api/(.*)$":
+      "<rootDir>/../../node_modules/@docknetwork/dock-blockchain-api/dist/esm/$1",
+    "^@docknetwork/dock-blockchain-api$":
+      "<rootDir>/../../node_modules/@docknetwork/dock-blockchain-api/dist/esm/index.js",
   },
   clearMocks: true,
   testTimeout: 30000,