+++ packages/cheqd-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/cheqd-blockchain-modules
 
+## 0.32.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.49.0
 
 ## 0.32.0
 
@@ -73,7 +79,7 @@
 - Updated dependencies
   - @docknetwork/credential-sdk@0.39.0
 
-## 0.28.1
+## 0.29.0
 
 ### Patch Changes