+++ scripts/migration/CHANGELOG.md
@@ -5,7 +5,9 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @docknetwork/dock-blockchain-api@0.23.0
+  - @docknetwork/cheqd-blockchain-api@0.35.0
+  - @docknetwork/credential-sdk@0.49.0
+  - @docknetwork/dock-blockchain-modules@0.29.0
 
 ## 0.9.1